src/tnn/datasets/nabirds.py
"""
NABirds dataset loader for image classification experiments
Placeholder implementation - will be completed when dataset is available
"""
import torch
from torch.utils.data import DataLoader, Dataset
from typing import Tuple, Optional
import os
from PIL import Image

from .transforms import get_nabirds_transforms

class NABirdsDataset(Dataset):
    """
    NABirds dataset implementation
    This is a placeholder that will be completed when the dataset is downloaded
    """
    
    def __init__(self, root_dir: str, train: bool = True, transform=None):
        self.root_dir = root_dir
        self.train = train
        self.transform = transform
        self.samples = []
        self.classes = []
        
        # This will be implemented once we have the dataset structure
        # Typical NABirds structure:
        # root_dir/
        #   images/
        #     001.Black_footed_Albatross/
        #       Black_Footed_Albatross_0001_796111.jpg
        #       ...
        #   train_test_split.txt
        #   classes.txt
        
        if os.path.exists(root_dir):
            self._load_dataset()
        else:
            print(f"Warning: NABirds dataset not found at {root_dir}")
            print("Please download the dataset and update the path")
            # Create dummy data for testing
            self._create_dummy_data()
    
    def _load_dataset(self):
        """Load actual NABirds dataset (to be implemented)"""
        # TODO: Implement actual dataset loading
        print("Loading NABirds dataset...")
        self._create_dummy_data()  # Placeholder
    
    def _create_dummy_data(self):
        """Create dummy data for testing when dataset is not available"""
        print("Creating dummy NABirds data for testing...")
        # Create 1000 dummy samples with 10 classes
        import torch
        for i in range(1000):
            class_id = i % 10
            self.samples.append((torch.randn(3, 224, 224), class_id))
        
        self.classes = [f"bird_class_{i}" for i in range(10)]
    
    def __len__(self):
        return len(self.samples)
    
    def __getitem__(self, idx):
        if isinstance(self.samples[idx], tuple):
            # Dummy data
            image, label = self.samples[idx]
            if self.transform:
                image = self.transform(image)
            return image, label
        else:
            # Real data (to be implemented)
            image_path, label = self.samples[idx]
            image = Image.open(image_path).convert('RGB')
            if self.transform:
                image = self.transform(image)
            return image, label


def get_nabirds_loaders(
    data_dir: str = './data/nabirds',
    batch_size: int = 64,
    frozen: bool = False,
    pretrained: bool = True,
    image_size: int = 224,
    num_workers: int = 4,
    pin_memory: bool = True
) -> Tuple[DataLoader, DataLoader, DataLoader]:
    """
    Get NABirds train, validation, and test data loaders
    
    Args:
        data_dir: Directory containing NABirds dataset
        batch_size: Batch size for data loaders
        frozen: Whether backbone is frozen (affects augmentation)
        pretrained: Whether using pretrained weights (affects normalization)
        image_size: Target image size (224 for ResNet)
        num_workers: Number of data loading workers
        pin_memory: Whether to pin memory for GPU
        
    Returns:
        Tuple of (train_loader, val_loader, test_loader)
    """
    
    # Get transforms
    train_transform, val_transform = get_nabirds_transforms(frozen, pretrained, image_size)
    
    # Load datasets
    train_dataset = NABirdsDataset(
        root_dir=data_dir,
        train=True,
        transform=train_transform
    )
    
    # For now, use the same dataset for validation and test
    # This will be updated when we have the actual train/test split
    val_dataset = NABirdsDataset(
        root_dir=data_dir,
        train=False,
        transform=val_transform
    )
    
    test_dataset = NABirdsDataset(
        root_dir=data_dir,
        train=False,
        transform=val_transform
    )
    
    # Create data loaders
    train_loader = DataLoader(
        train_dataset,
        batch_size=batch_size,
        shuffle=True,
        num_workers=num_workers,
        pin_memory=pin_memory
    )
    
    val_loader = DataLoader(
        val_dataset,
        batch_size=batch_size,
        shuffle=False,
        num_workers=num_workers,
        pin_memory=pin_memory
    )
    
    test_loader = DataLoader(
        test_dataset,
        batch_size=batch_size,
        shuffle=False,
        num_workers=num_workers,
        pin_memory=pin_memory
    )
    
    print(f"NABirds Dataset loaded:")
    print(f"  Train: {len(train_dataset)} samples")
    print(f"  Val: {len(val_dataset)} samples")
    print(f"  Test: {len(test_dataset)} samples")
    print(f"  Batch size: {batch_size}")
    print(f"  Image size: {image_size}x{image_size}")
    print(f"  Frozen backbone: {frozen}")
    print(f"  Pretrained: {pretrained}")
    
    return train_loader, val_loader, test_loader
